Document the vr-threading crate
